requirements.txt
yt-dlp>=2023.12.30
openai-whisper
tqdm
python-dotenv
openai
pyyaml
youtube-transcript-api
anthropic
